Basketball teams' portal
Designed and implemented a web interface for keeping track of a youth
basketball conference. Each player or coach will be able to look at the
team information, including what games have been played and who won.
Additionally, there is an administrative portal where new teams,
coaches, players, and games can be added as well as a way for
referees to enter the points each team made during a particular game.